requirements.txt
absl-py==2.3.1
altair==4.2.2
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
entrypoints==0.4
flatbuffers==25.2.10
fonttools==4.59.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.6.2
mdurl==0.1.2
numpy==1.23.5
oauthlib==3.3.1
opencv-python==4.6.0.66
opt_einsum==3.4.0
packaging==25.0
pandas==1.5.1
pillow==11.3.0
protobuf==3.19.6
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydeck==0.9.1
Pygments==2.19.2
Pympler==1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
scikit-learn==1.1.3
scipy==1.13.1
six==1.17.0
smmap==5.0.2
streamlit==1.22.0
tenacity==8.5.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
toolz==1.0.0
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.23.0
